+++ demo/src/index.js
@@ -12,7 +12,20 @@ const MarkerIcon = leaflet.icon({
     popupAnchor: [0, 0],
 });
 class Demo extends React.Component {
+    state = {
+        init: true
+    };
+    componentDidMount(){
+        setTimeout(()=>{
+            console.log('this.init');
+            this.setState({
+                init: false
+            })
+        }, 3000);
+    }
     render() {
+        let lay = <LeafletPolyline positions={[[45.9205, 0.4394], [45.9205, 0.9228], [45.9205, 2.5488], [45.9205, 3.3837]]}/>;
+
         return <div style={{height: 500, width: '100%'}}>
             <h1>React-Leaflet-Draw example</h1>
             <Map center={[45.9205, 0.9228]} zoom={9} zoomControl={false} style={{height: 500, width: '100%'}}>
@@ -21,13 +34,13 @@ class Demo extends React.Component {
                     url="http://{s}.tile.osm.org/{z}/{x}/{y}.png"
                 />
 
-                <Marker position={[45.9205, 0.4394]} icon={MarkerIcon}>
-                </Marker>
+                {this.state.init && <Marker position={[45.9205, 0.4394]} icon={MarkerIcon}>
+                </Marker>}
 
-                <Marker position={[45.9205, 3.3837]} icon={MarkerIcon}>
-                </Marker>
+                {!this.state.init && <Marker position={[45.9205, 3.3837]} icon={MarkerIcon}>
+                </Marker>}
 
-                <LeafletPolyline positions={[[45.9205, 0.4394], [45.9205, 0.9228], [45.9205, 2.5488], [45.9205, 3.3837]]}/>
+                {this.state.init && lay}
             </Map>
         </div>
     }