+++ CHANGELOG.md
@@ -6,5 +6,17 @@ git Project:
 - [bug Reporting](http://gitlab.peykasa.ir/react/components/leaflet-polyline/issues)
 
 
+ Leaflet Polyline v1.0.2
+----------------------------------
+- use layerGroup
+- remove able layers
+
+
+ Leaflet Polyline v1.0.1
+----------------------------------
+- fix bugs
+
+
  Leaflet Polyline v1.0.0
 ----------------------------------
+- first release